requirements.txt
asyncpg==0.29.0
pydantic==2.8.2
pyyaml==6.0.2
python-dotenv==1.0.1
click==8.1.7
loguru==0.7.2
prometheus-client==0.20.0
python-telegram-bot==21.6
aiohttp==3.10.5
uvloop==0.19.0 ; sys_platform == 'linux'
pytest==8.3.3
pandas==2.2.2
pip install nest_asyncio==1.6.0